+++ .storybook/main.js
@@ -1,24 +1,20 @@
 module.exports = {
-  stories: [
-    // files MUST end with .story.mdx or .stories.mdx (@see https://github.com/storybookjs/storybook/issues/9918)
-    "../src/themes/*.story.mdx",
-    "../src/components/**/*.story.mdx",
-    "../src/components/**/*.stories.mdx",
-    "../src/snowflakes/**/*.story.mdx",
-    "../src/snowflakes/**/*.stories.mdx",
+  framework: "@storybook/web-components",
+  core: {
+    builder: "webpack5"
+  },
+  staticDirs: [
+    "../dist",
+    "../src/assets",
+    { from: "../src/assets/favicon.png", to: "/images/favicon.png" }
   ],
+  stories: ["../**/*.stories.mdx"],
   addons: [
-    "@storybook/addon-viewport/register",
-    "@storybook/addon-knobs/register",
+    "storybook-addon-stencil",
+    "@storybook/addon-docs",
+    "@storybook/addon-controls",
+    "@storybook/addon-viewport",
+    "@storybook/addon-knobs",
     "@storybook/addon-a11y",
-    "storybook-addons-abstract",
-    {
-      name: "@storybook/addon-docs",
-      options: {
-        configureJSX: true,
-        babelOptions: {},
-        sourceLoaderOptions: null,
-      },
-    },
   ],
 };