+++ src/components.d.ts
@@ -6,7 +6,7 @@
  */
 import { HTMLStencilElement, JSXBase } from "@stencil/core/internal";
 import { AlertTypes, LicenseTypeEnum, MenuItem as MenuItem1, TooltipPosition } from "./beans/index";
-import { AvatarSize, ButtonSizeEnum, ButtonVariantBean, CardVariants, ComboItemBean, DictionaryData, DividerOrientation, DividerSize, ExpandableListButtonAlign, ExpandableListStyle, HeaderUserData, InputStatusBean, InputTypeBean, ListDividerType, ListSize, MenuItem, NotificationType, PocketStatus, PopoverBorderRadius, PopoverPosition, PopoverShadow, SelectItemBean, Size, SortDirection, StatusTagStatus, TabOrientationBean, TabSizeBean, ThemeVariant, ThemeVariantBean, ToastNotificationPositionsTypes, ToastNotificationTransitionTypes, ToastNotificationTypes, ZChipType, ZTableRowExpandedType } from "./beans";
+import { AvatarSize, ButtonSizeEnum, ButtonVariantBean, ButtonVariantEnum, CardVariants, ComboItemBean, DictionaryData, DividerOrientation, DividerSize, ExpandableListButtonAlign, ExpandableListStyle, HeaderUserData, InputStatusBean, InputTypeBean, ListDividerType, ListSize, MenuItem, NotificationType, OffCanvasVariantsEnum, PocketStatus, PopoverBorderRadius, PopoverPosition, PopoverShadow, SelectItemBean, Size, SortDirection, StatusTagStatus, TabOrientationBean, TabSizeBean, ThemeVariant, ThemeVariantBean, ToastNotificationPositionsTypes, ToastNotificationTransitionTypes, ToastNotificationTypes, TransitionDirectionEnum, ZChipType, ZFileUploadTypeEnum, ZSectionTitleDividerPosition, ZTableRowExpandedType, ZtoggleSwitchPositionEnum } from "./beans";
 import { ListItemBean } from "./beans/index.js";
 import { ZTypographyLevels } from "./components/typography/z-typography/index";
 export namespace Components {
@@ -333,6 +333,37 @@ export namespace Components {
           * [optional] Divider size
          */
         "size"?: DividerSize;
+    }
+    interface ZDragdropArea {
+    }
+    interface ZFile {
+        "fileNumber": number;
+    }
+    interface ZFileUpload {
+        /**
+          * Prop indicating the accepted file type: ex ".pdf, .doc, .jpg"
+         */
+        "acceptedFormat"?: string;
+        /**
+          * Prop indicating the button variant
+         */
+        "buttonVariant"?: ButtonVariantEnum;
+        /**
+          * Description
+         */
+        "description"?: string;
+        /**
+          * Max file dimension in Megabyte
+         */
+        "fileMaxSize"?: number;
+        /**
+          * Title
+         */
+        "mainTitle"?: string;
+        /**
+          * Prop indicating the file upload type - can be default or dragdrop
+         */
+        "type": ZFileUploadTypeEnum;
     }
     interface ZFooter {
         /**
@@ -1037,25 +1068,44 @@ export namespace Components {
     }
     interface ZNotification {
         /**
-          * action button text
+          * Action button text
          */
         "actiontext"?: string;
         /**
-          * icon on the left of the content
+          * Name of the icon on the left of the content
          */
         "contenticonname"?: string;
         /**
-          * enable close icon
+          * Enable close icon
          */
         "showclose"?: boolean;
         /**
-          * enable shadow
+          * Enable shadow.
+          * @deprecated shadow is available only for the `sticky` version of the notification.
          */
         "showshadow"?: boolean;
         /**
-          * alert variant type
+          * Enable sticky notification bar.
+         */
+        "sticky"?: boolean;
+        /**
+          * Alert variant type
          */
         "type": NotificationType;
+    }
+    interface ZOffcanvas {
+        /**
+          * open component. Default: false
+         */
+        "open": boolean;
+        /**
+          * open content transitioning in a specified direction left | right. Default: left
+         */
+        "transitiondirection"?: TransitionDirectionEnum;
+        /**
+          * Offcanvas variant. Can be one of "overlay", "pushcontent" Default variant: pushcontent
+         */
+        "variant"?: OffCanvasVariantsEnum;
     }
     interface ZOtp {
         "inputNum"?: number;
@@ -1213,6 +1263,16 @@ export namespace Components {
           * [optional] Show or hide arrow
          */
         "showArrow"?: boolean;
+    }
+    interface ZSectionTitle {
+        /**
+          * Divider position for the primary title. This prop only works if the secondary title is not set.
+         */
+        "dividerPosition": ZSectionTitleDividerPosition;
+        /**
+          * Whether the primary title text is uppercase.
+         */
+        "uppercase": boolean;
     }
     interface ZSelect {
         /**
@@ -1519,6 +1579,12 @@ export namespace Components {
          */
         "opened"?: boolean;
     }
+    interface ZToggleSwitch {
+        "checked"?: boolean;
+        "disabled"?: boolean;
+        "htmlid": string;
+        "labelPosition"?: ZtoggleSwitchPositionEnum;
+    }
     interface ZTooltip {
         /**
           * The selector or the element bound with the tooltip.
@@ -1682,6 +1748,24 @@ declare global {
     var HTMLZDividerElement: {
         prototype: HTMLZDividerElement;
         new (): HTMLZDividerElement;
+    };
+    interface HTMLZDragdropAreaElement extends Components.ZDragdropArea, HTMLStencilElement {
+    }
+    var HTMLZDragdropAreaElement: {
+        prototype: HTMLZDragdropAreaElement;
+        new (): HTMLZDragdropAreaElement;
+    };
+    interface HTMLZFileElement extends Components.ZFile, HTMLStencilElement {
+    }
+    var HTMLZFileElement: {
+        prototype: HTMLZFileElement;
+        new (): HTMLZFileElement;
+    };
+    interface HTMLZFileUploadElement extends Components.ZFileUpload, HTMLStencilElement {
+    }
+    var HTMLZFileUploadElement: {
+        prototype: HTMLZFileUploadElement;
+        new (): HTMLZFileUploadElement;
     };
     interface HTMLZFooterElement extends Components.ZFooter, HTMLStencilElement {
     }
@@ -1923,6 +2007,12 @@ declare global {
         prototype: HTMLZNotificationElement;
         new (): HTMLZNotificationElement;
     };
+    interface HTMLZOffcanvasElement extends Components.ZOffcanvas, HTMLStencilElement {
+    }
+    var HTMLZOffcanvasElement: {
+        prototype: HTMLZOffcanvasElement;
+        new (): HTMLZOffcanvasElement;
+    };
     interface HTMLZOtpElement extends Components.ZOtp, HTMLStencilElement {
     }
     var HTMLZOtpElement: {
@@ -1976,6 +2066,12 @@ declare global {
     var HTMLZPopoverElement: {
         prototype: HTMLZPopoverElement;
         new (): HTMLZPopoverElement;
+    };
+    interface HTMLZSectionTitleElement extends Components.ZSectionTitle, HTMLStencilElement {
+    }
+    var HTMLZSectionTitleElement: {
+        prototype: HTMLZSectionTitleElement;
+        new (): HTMLZSectionTitleElement;
     };
     interface HTMLZSelectElement extends Components.ZSelect, HTMLStencilElement {
     }
@@ -2091,6 +2187,12 @@ declare global {
         prototype: HTMLZToggleButtonElement;
         new (): HTMLZToggleButtonElement;
     };
+    interface HTMLZToggleSwitchElement extends Components.ZToggleSwitch, HTMLStencilElement {
+    }
+    var HTMLZToggleSwitchElement: {
+        prototype: HTMLZToggleSwitchElement;
+        new (): HTMLZToggleSwitchElement;
+    };
     interface HTMLZTooltipElement extends Components.ZTooltip, HTMLStencilElement {
     }
     var HTMLZTooltipElement: {
@@ -2135,6 +2237,9 @@ declare global {
         "z-cookiebar": HTMLZCookiebarElement;
         "z-date-picker": HTMLZDatePickerElement;
         "z-divider": HTMLZDividerElement;
+        "z-dragdrop-area": HTMLZDragdropAreaElement;
+        "z-file": HTMLZFileElement;
+        "z-file-upload": HTMLZFileUploadElement;
         "z-footer": HTMLZFooterElement;
         "z-footer-link": HTMLZFooterLinkElement;
         "z-footer-section": HTMLZFooterSectionElement;
@@ -2175,6 +2280,7 @@ declare global {
         "z-navigation-tab": HTMLZNavigationTabElement;
         "z-navigation-tabs": HTMLZNavigationTabsElement;
         "z-notification": HTMLZNotificationElement;
+        "z-offcanvas": HTMLZOffcanvasElement;
         "z-otp": HTMLZOtpElement;
         "z-pagination-bar": HTMLZPaginationBarElement;
         "z-pagination-page": HTMLZPaginationPageElement;
@@ -2184,6 +2290,7 @@ declare global {
         "z-pocket-header": HTMLZPocketHeaderElement;
         "z-pocket-message": HTMLZPocketMessageElement;
         "z-popover": HTMLZPopoverElement;
+        "z-section-title": HTMLZSectionTitleElement;
         "z-select": HTMLZSelectElement;
         "z-slideshow": HTMLZSlideshowElement;
         "z-status-tag": HTMLZStatusTagElement;
@@ -2203,6 +2310,7 @@ declare global {
         "z-toast-notification": HTMLZToastNotificationElement;
         "z-toast-notification-list": HTMLZToastNotificationListElement;
         "z-toggle-button": HTMLZToggleButtonElement;
+        "z-toggle-switch": HTMLZToggleSwitchElement;
         "z-tooltip": HTMLZTooltipElement;
         "z-typography": HTMLZTypographyElement;
         "z-user-dropdown": HTMLZUserDropdownElement;
@@ -2566,6 +2674,49 @@ declare namespace LocalJSX {
           * [optional] Divider size
          */
         "size"?: DividerSize;
+    }
+    interface ZDragdropArea {
+        /**
+          * Emitted when user drop one or more files
+         */
+        "onFileDropped"?: (event: CustomEvent<any>) => void;
+    }
+    interface ZFile {
+        "fileNumber"?: number;
+        /**
+          * Emitted when a z-file component is removed from the DOM
+         */
+        "onRemoveFile"?: (event: CustomEvent<any>) => void;
+    }
+    interface ZFileUpload {
+        /**
+          * Prop indicating the accepted file type: ex ".pdf, .doc, .jpg"
+         */
+        "acceptedFormat"?: string;
+        /**
+          * Prop indicating the button variant
+         */
+        "buttonVariant"?: ButtonVariantEnum;
+        /**
+          * Description
+         */
+        "description"?: string;
+        /**
+          * Max file dimension in Megabyte
+         */
+        "fileMaxSize"?: number;
+        /**
+          * Title
+         */
+        "mainTitle"?: string;
+        /**
+          * Emitted when user select one or more files
+         */
+        "onFileInput"?: (event: CustomEvent<any>) => void;
+        /**
+          * Prop indicating the file upload type - can be default or dragdrop
+         */
+        "type"?: ZFileUploadTypeEnum;
     }
     interface ZFooter {
         /**
@@ -3367,33 +3518,56 @@ declare namespace LocalJSX {
     }
     interface ZNotification {
         /**
-          * action button text
+          * Action button text
          */
         "actiontext"?: string;
         /**
-          * icon on the left of the content
+          * Name of the icon on the left of the content
          */
         "contenticonname"?: string;
         /**
-          * notification action event
+          * Call to action clicked
          */
         "onNotificationAction"?: (event: CustomEvent<any>) => void;
         /**
-          * notification close event
+          * Close button clicked
          */
         "onNotificationClose"?: (event: CustomEvent<any>) => void;
         /**
-          * enable close icon
+          * Enable close icon
          */
         "showclose"?: boolean;
         /**
-          * enable shadow
+          * Enable shadow.
+          * @deprecated shadow is available only for the `sticky` version of the notification.
          */
         "showshadow"?: boolean;
         /**
-          * alert variant type
+          * Enable sticky notification bar.
+         */
+        "sticky"?: boolean;
+        /**
+          * Alert variant type
          */
         "type"?: NotificationType;
+    }
+    interface ZOffcanvas {
+        /**
+          * emitted when open prop changes
+         */
+        "onCanvasOpenStatusChanged"?: (event: CustomEvent<any>) => void;
+        /**
+          * open component. Default: false
+         */
+        "open"?: boolean;
+        /**
+          * open content transitioning in a specified direction left | right. Default: left
+         */
+        "transitiondirection"?: TransitionDirectionEnum;
+        /**
+          * Offcanvas variant. Can be one of "overlay", "pushcontent" Default variant: pushcontent
+         */
+        "variant"?: OffCanvasVariantsEnum;
     }
     interface ZOtp {
         "inputNum"?: number;
@@ -3568,6 +3742,16 @@ declare namespace LocalJSX {
           * [optional] Show or hide arrow
          */
         "showArrow"?: boolean;
+    }
+    interface ZSectionTitle {
+        /**
+          * Divider position for the primary title. This prop only works if the secondary title is not set.
+         */
+        "dividerPosition"?: ZSectionTitleDividerPosition;
+        /**
+          * Whether the primary title text is uppercase.
+         */
+        "uppercase"?: boolean;
     }
     interface ZSelect {
         /**
@@ -3887,6 +4071,13 @@ declare namespace LocalJSX {
          */
         "opened"?: boolean;
     }
+    interface ZToggleSwitch {
+        "checked"?: boolean;
+        "disabled"?: boolean;
+        "htmlid"?: string;
+        "labelPosition"?: ZtoggleSwitchPositionEnum;
+        "onToggleClick"?: (event: CustomEvent<any>) => void;
+    }
     interface ZTooltip {
         /**
           * The selector or the element bound with the tooltip.
@@ -3967,6 +4158,9 @@ declare namespace LocalJSX {
         "z-cookiebar": ZCookiebar;
         "z-date-picker": ZDatePicker;
         "z-divider": ZDivider;
+        "z-dragdrop-area": ZDragdropArea;
+        "z-file": ZFile;
+        "z-file-upload": ZFileUpload;
         "z-footer": ZFooter;
         "z-footer-link": ZFooterLink;
         "z-footer-section": ZFooterSection;
@@ -4007,6 +4201,7 @@ declare namespace LocalJSX {
         "z-navigation-tab": ZNavigationTab;
         "z-navigation-tabs": ZNavigationTabs;
         "z-notification": ZNotification;
+        "z-offcanvas": ZOffcanvas;
         "z-otp": ZOtp;
         "z-pagination-bar": ZPaginationBar;
         "z-pagination-page": ZPaginationPage;
@@ -4016,6 +4211,7 @@ declare namespace LocalJSX {
         "z-pocket-header": ZPocketHeader;
         "z-pocket-message": ZPocketMessage;
         "z-popover": ZPopover;
+        "z-section-title": ZSectionTitle;
         "z-select": ZSelect;
         "z-slideshow": ZSlideshow;
         "z-status-tag": ZStatusTag;
@@ -4035,6 +4231,7 @@ declare namespace LocalJSX {
         "z-toast-notification": ZToastNotification;
         "z-toast-notification-list": ZToastNotificationList;
         "z-toggle-button": ZToggleButton;
+        "z-toggle-switch": ZToggleSwitch;
         "z-tooltip": ZTooltip;
         "z-typography": ZTypography;
         "z-user-dropdown": ZUserDropdown;
@@ -4064,6 +4261,9 @@ declare module "@stencil/core" {
             "z-cookiebar": LocalJSX.ZCookiebar & JSXBase.HTMLAttributes<HTMLZCookiebarElement>;
             "z-date-picker": LocalJSX.ZDatePicker & JSXBase.HTMLAttributes<HTMLZDatePickerElement>;
             "z-divider": LocalJSX.ZDivider & JSXBase.HTMLAttributes<HTMLZDividerElement>;
+            "z-dragdrop-area": LocalJSX.ZDragdropArea & JSXBase.HTMLAttributes<HTMLZDragdropAreaElement>;
+            "z-file": LocalJSX.ZFile & JSXBase.HTMLAttributes<HTMLZFileElement>;
+            "z-file-upload": LocalJSX.ZFileUpload & JSXBase.HTMLAttributes<HTMLZFileUploadElement>;
             "z-footer": LocalJSX.ZFooter & JSXBase.HTMLAttributes<HTMLZFooterElement>;
             "z-footer-link": LocalJSX.ZFooterLink & JSXBase.HTMLAttributes<HTMLZFooterLinkElement>;
             "z-footer-section": LocalJSX.ZFooterSection & JSXBase.HTMLAttributes<HTMLZFooterSectionElement>;
@@ -4104,6 +4304,7 @@ declare module "@stencil/core" {
             "z-navigation-tab": LocalJSX.ZNavigationTab & JSXBase.HTMLAttributes<HTMLZNavigationTabElement>;
             "z-navigation-tabs": LocalJSX.ZNavigationTabs & JSXBase.HTMLAttributes<HTMLZNavigationTabsElement>;
             "z-notification": LocalJSX.ZNotification & JSXBase.HTMLAttributes<HTMLZNotificationElement>;
+            "z-offcanvas": LocalJSX.ZOffcanvas & JSXBase.HTMLAttributes<HTMLZOffcanvasElement>;
             "z-otp": LocalJSX.ZOtp & JSXBase.HTMLAttributes<HTMLZOtpElement>;
             "z-pagination-bar": LocalJSX.ZPaginationBar & JSXBase.HTMLAttributes<HTMLZPaginationBarElement>;
             "z-pagination-page": LocalJSX.ZPaginationPage & JSXBase.HTMLAttributes<HTMLZPaginationPageElement>;
@@ -4113,6 +4314,7 @@ declare module "@stencil/core" {
             "z-pocket-header": LocalJSX.ZPocketHeader & JSXBase.HTMLAttributes<HTMLZPocketHeaderElement>;
             "z-pocket-message": LocalJSX.ZPocketMessage & JSXBase.HTMLAttributes<HTMLZPocketMessageElement>;
             "z-popover": LocalJSX.ZPopover & JSXBase.HTMLAttributes<HTMLZPopoverElement>;
+            "z-section-title": LocalJSX.ZSectionTitle & JSXBase.HTMLAttributes<HTMLZSectionTitleElement>;
             "z-select": LocalJSX.ZSelect & JSXBase.HTMLAttributes<HTMLZSelectElement>;
             "z-slideshow": LocalJSX.ZSlideshow & JSXBase.HTMLAttributes<HTMLZSlideshowElement>;
             "z-status-tag": LocalJSX.ZStatusTag & JSXBase.HTMLAttributes<HTMLZStatusTagElement>;
@@ -4132,6 +4334,7 @@ declare module "@stencil/core" {
             "z-toast-notification": LocalJSX.ZToastNotification & JSXBase.HTMLAttributes<HTMLZToastNotificationElement>;
             "z-toast-notification-list": LocalJSX.ZToastNotificationList & JSXBase.HTMLAttributes<HTMLZToastNotificationListElement>;
             "z-toggle-button": LocalJSX.ZToggleButton & JSXBase.HTMLAttributes<HTMLZToggleButtonElement>;
+            "z-toggle-switch": LocalJSX.ZToggleSwitch & JSXBase.HTMLAttributes<HTMLZToggleSwitchElement>;
             "z-tooltip": LocalJSX.ZTooltip & JSXBase.HTMLAttributes<HTMLZTooltipElement>;
             "z-typography": LocalJSX.ZTypography & JSXBase.HTMLAttributes<HTMLZTypographyElement>;
             "z-user-dropdown": LocalJSX.ZUserDropdown & JSXBase.HTMLAttributes<HTMLZUserDropdownElement>;