+++ src/components/notification/z-notification/index.spec.ts
@@ -3,8 +3,7 @@ import { newSpecPage } from "@stencil/core/testing";
 import { ZNotification } from "./index";
 
 describe("Suite test ZNotification", () => {
-  it("Test render ZNotification with icon, action and close icon", async () => {
-
+  it("Test render ZNotification with icon, action and close button", async () => {
     const page = await newSpecPage({
       components: [ZNotification],
       html: `<z-notification
@@ -12,69 +11,54 @@ describe("Suite test ZNotification", () => {
         actiontext="Annulla"
         type="success"
         showclose="true"
-      >Io ne ho viste cose</z-notification>`
+      >Testo della notifica</z-notification>`
     });
 
     expect(page.root).toEqualHtml(`
       <z-notification actiontext="Annulla" contenticonname="gear" showclose="true" type="success">
         <mock:shadow-root>
-          <div class="notification-container success-notification">
+          <z-icon
+            class="status-icon"
+            name="gear"
+            width="16"
+            height="16"
+          ></z-icon>
+          <div class="content-container">
+            <div class="content-text"><slot></slot></div>
+            <button class="action-button" type="button">Annulla</button>
+          </div>
+          <button class="close-button" type="button">
             <z-icon
-              name="gear"
-              width="16"
-              height="16"
-            ></z-icon>
-            <div class="content-container">
-              <z-body class="content-text" level="4"><slot></slot></z-body>
-                  <z-body
-                    class="action-text"
-                    role="button"
-                    tabindex="0"
-                    level="5"
-                    variant="semibold"
-                  >
-                    Annulla
-                  </z-body>
-            </div>
-            <z-icon
-              class="close-icon"
               name="multiply-circle"
               width="16"
               height="16"
             />
-
-          </div>
+          </button>
         </mock:shadow-root>
-        Io ne ho viste cose
+        Testo della notifica
       </z-notification>
     `);
   });
 
-  it("Test render ZNotification without icon, action and close icon", async () => {
-
+  it("Test render ZNotification without icon, action and close button", async () => {
     const page = await newSpecPage({
       components: [ZNotification],
-      html: `<z-notification
-        type="success"
-      >Io ne ho viste cose</z-notification>`
+      html: `<z-notification type="success">Testo della notifica</z-notification>`
     });
 
     expect(page.root).toEqualHtml(`
       <z-notification type="success">
         <mock:shadow-root>
-          <div class="notification-container success-notification">
-            <div class="content-container">
-              <z-body class="content-text" level="4"><slot></slot></z-body>
-            </div>
+          <div class="content-container">
+            <div class="content-text"><slot></slot></div>
           </div>
         </mock:shadow-root>
-        Io ne ho viste cose
+        Testo della notifica
       </z-notification>
     `);
   });
 
   it("Test render ZNotification with icon, action", async () => {
-
     const page = await newSpecPage({
       components: [ZNotification],
       html: `<z-notification
@@ -82,129 +66,24 @@ describe("Suite test ZNotification", () => {
         actiontext="Annulla"
         type="success"
         showclose="false"
-      >Io ne ho viste cose</z-notification>`
+      >Testo della notifica</z-notification>`
     });
 
     expect(page.root).toEqualHtml(`
       <z-notification actiontext="Annulla" contenticonname="gear" showclose="false" type="success">
         <mock:shadow-root>
-          <div class="notification-container success-notification">
-            <z-icon
-              name="gear"
-              width="16"
-              height="16"
-            ></z-icon>
-            <div class="content-container">
-              <z-body class="content-text" level="4"><slot></slot></z-body>
-                  <z-body
-                    class="action-text"
-                    role="button"
-                    tabindex="0"
-                    level="5"
-                    variant="semibold"
-                  >
-                    Annulla
-                  </z-body>
-            </div>
+          <z-icon
+            class="status-icon"
+            name="gear"
+            width="16"
+            height="16"
+          ></z-icon>
+          <div class="content-container">
+            <div class="content-text"><slot></slot></div>
+            <button class="action-button" type="button">Annulla</button>
           </div>
         </mock:shadow-root>
-        Io ne ho viste cose
-      </z-notification>
-    `);
-  });
-
-  it("renders ZNotification with shadow", async () => {
-    const page = await newSpecPage({
-      components: [ZNotification],
-      html: `<z-notification
-        contenticonname="gear"
-        actiontext="Annulla"
-        type="success"
-        showclose="false"
-        showshadow="true"
-      >Io ne ho viste cose</z-notification>`
-    });
-
-    expect(page.root).toEqualHtml(`
-      <z-notification actiontext="Annulla" contenticonname="gear" showclose="false" showshadow="true" type="success">
-        <mock:shadow-root>
-          <div class="notification-container success-notification shadow">
-            <z-icon
-              name="gear"
-              width="16"
-              height="16"
-            ></z-icon>
-            <div class="content-container">
-              <z-body class="content-text" level="4"><slot></slot></z-body>
-                  <z-body
-                    class="action-text"
-                    role="button"
-                    tabindex="0"
-                    level="5"
-                    variant="semibold"
-                  >
-                    Annulla
-                  </z-body>
-            </div>
-          </div>
-        </mock:shadow-root>
-        Io ne ho viste cose
-      </z-notification>
-    `);
-  });
-
-  it("renders ZNotification with warning type", async () => {
-    const page = await newSpecPage({
-      components: [ZNotification],
-      html: `<z-notification
-        contenticonname="gear"
-        type="warning"
-      >Io ne ho viste cose</z-notification>`
-    });
-
-    expect(page.root).toEqualHtml(`
-      <z-notification contenticonname="gear" type="warning">
-        <mock:shadow-root>
-          <div class="notification-container warning-notification">
-            <z-icon
-              name="gear"
-              width="16"
-              height="16"
-            ></z-icon>
-            <div class="content-container">
-              <z-body class="content-text" level="4"><slot></slot></z-body>
-            </div>
-          </div>
-        </mock:shadow-root>
-        Io ne ho viste cose
-      </z-notification>
-    `);
-  });
-
-  it("renders ZNotification with error type", async () => {
-    const page = await newSpecPage({
-      components: [ZNotification],
-      html: `<z-notification
-        contenticonname="gear"
-        type="error"
-      >Io ne ho viste cose</z-notification>`
-    });
-
-    expect(page.root).toEqualHtml(`
-      <z-notification contenticonname="gear" type="error">
-        <mock:shadow-root>
-          <div class="notification-container error-notification">
-            <z-icon
-              name="gear"
-              width="16"
-              height="16"
-            ></z-icon>
-            <div class="content-container">
-              <z-body class="content-text" level="4"><slot></slot></z-body>
-            </div>
-          </div>
-        </mock:shadow-root>
-        Io ne ho viste cose
+        Testo della notifica
       </z-notification>
     `);
   });